extract basic-auth and external jwt-auth into their own files
